backend/requirements.txt
asgiref==3.3.2
Django==3.2
django-extra-fields==3.0.2
django-filter==22.1
djangorestframework
djangorestframework-simplejwt
djoser==2.1.0
psycopg2-binary
pycodestyle==2.9.1
PyJWT==2.4.0
pytz==2020.1
sqlparse==0.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
gunicorn==20.1.0
Pillow
django-colorfield